requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
numpy==1.23.4
opencv-python==4.6.0.66
python-dotenv==0.21.0
requests==2.28.1
telebot==0.0.4
urllib3==1.26.12
